Reworking Dashboard Components
