+++ Cargo.toml
@@ -16,6 +16,7 @@ getrandom = { version = "0.2", features = ["js", "wasm-bindgen"] }
 
 strum = "0.24"
 strum_macros = "0.24"
+egui_plot = "0.30.0"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 env_logger = "0.11"
