+++ src/i18n/en.yml
@@ -35,10 +35,7 @@ app:
       choose: 'Choose a font size'
   dashboard:
     title: 'Dashboard'
-    metrics:
-      total_jobs: 'Total Jobs'
-      running: 'Running'
-      waiting: 'Waiting'
+    total_jobs: 'Total Jobs'
   gantt:
     title: 'Gantt Chart - OAR Jobs'
     zoom_in: 'Zoom +'
