+++ src/views/components/dashboard_components/mod.rs
@@ -1,0 +1,4 @@
+pub mod job_table;
+pub mod metric_box;
+pub mod metric_grid;
+pub mod metric_chart;
