+++ src/views/components/mod.rs
@@ -1,4 +1,3 @@
-pub mod metric_box;
-pub mod job_table;
 pub mod time_selector;
 pub mod job_details;
+pub mod dashboard_components;
